Allow a directory to be replaced with a file
Consider a target directory containing a file `dir/file` and a source
directory containing a file `dir`. Note that in the target `dir` is a directory
and in the source `dir` is file. Currently `kudusync` does not remove the
directory in the target folder before attempting to create the file `dir`.

The fix is to delete unneeded directories first before starting to copy
files.

There is a test case in https://github.com/projectkudu/KuduSync/pull/40
It fails on the master branch of this repo and passes with this PR.